Change configuration build path to be absolute
While specifying the build path relative to something might be a useful
idea in principle, we would need a clearer story for how it is specified
and resolved. When passing a path to a command-line argument, it should
be resolved relative to the CWD, not to the package root by default.
Incidentally, it doesn't really make sense to have a global default,
since different build systems are likely to want different defaults. For
now, simplify this configuration to an optional absolute path (passing a
relative path on the command line will resolve relative to the CWD).
Each build system can interpret `nil` as its preferred default.