[swiftpm] Statically link to libSwiftPM
Switch to using the swiftpm library with automatic linkage, which in
practice means we will statically link to libSwiftPM using the standard
build. This change allows us to install the sourcekit-lsp binary without
also installing a copy of libSwiftPM.dylib (and since swiftpm isn't
using a stable ABI we couldn't share that with anything). Previously
this was blocked by conflicts between the llvm code in llbuild and in
indexstore-db, but this has been resolved.