Remove Package.resolved and move to HEAD
Always use the latest indexstore-db and swiftpm from the corresponding
branch (currently master). This is important for swiftpm, because it
does not *guarantee* stable behaviour, so we need to match the version.
It also makes sourcekit-lsp behave more like the rest of the swift
toolchain projects, who all live on HEAD for their dependencies.