[buildsystem] Introduce BuildSystemManager that handles main-file lookup
The build system manager will integrate with the indexer to provide
header to main file mappings, including keeping those mappings
up-to-date as `#include`s are added and removed. This allows us to
provide accurate compiler arguments for a header based on a real
translation unit. When there are no main files found, we continue to
fallback to asking the build system.

The build system manager also gives us a place to put a client-side
setting cache, since multiple headers may map to a single main file.

For now, the BuildSystemManager is only used by tests. A subsequent
commit will wire it up to the index and server.