Dup and redirect stdout to defend against accidental print()
We currently have a regression where sourcekit-lsp stops working  in
some packages, because swiftpm started printing a warning to stdout.
While we should also fix swiftpm, it behooves us to be more robust in
sourcekit-lsp in case this happens again. Dup stdout for use by the
protocol, and redirect the stdout file descriptor to stderr, which does
not interfere with the protocol.