SKSwiftPMWorkspace: make `testBasicCXXArgs` pass on Windows
We would previously fail to correctly build the GNU dependency outputs
on Windows since we were checking the file system representation of the
path against the internal path representation.  This fixes the
generation of that command as well as the additional checks in the test
case to allow the test case to succeed on Windows.