Fix a nondeterministic test failure in sourcekitd crash recovery
In some cases (especially if the system is under heavy load),
sourcekitd's crash is scheduled while the document update is in progress.
In this case, the SwiftLanguageServer's queue is blocked until the
document update is handled, which only happens once semantic editor
functionality has been restored.
Because the queue is blocked waiting for the document update to be
handled, we don't receive the crash notification within 5 seconds.
Increasing the timeout to 15 seconds to allow for 10 seconds semantic
functionality delay fixes the issue.

While I haven't been able to reproduce rdar://73717447, I suspect a
similar issue occurs for clang crash recovery and increasing the timeout
there should also resolve the issue.

This commit also unifies the two async blocks in SwiftLanguageServer's
notification handling into one. While it isn't strictly related to the
issue described above, there should be less room for errors if we
disallow scheduling of other blocks on the queue in between a
notification handling and document update handling.