Fix a nondeterministic sourcektid crash recovery test failure
In some cases (especially if the system is under heavy load), we crash
sourcekitd before having handled the documentupdate we received from
either the open request. If we then perform the 0,0 replacetext
synchronously in handleDocumentUpdate to receive the semantic
diagnostics, sourcekitd blocks until it has restored semantic
functionality and thus the LSP server's queue is blocked for that entire
duration.

To avoid the data race, make sure the document is fully opened before
crashing sourcekitd.