Add comment in Development.md that updating `PATH` is important
I’ve spend an hour trying to figure out why the test suite didn’t pass, because I was trying to be clever and not modify my `PATH` to have `<path_to_swift_toolchain>/usr/bin` as the first entry. So: Add a comment, telling you to not try and be clever.