vscodeignore: update the ignore rules
Update the ignore rules to reduce the package size and remove unnecesary
files.  With the new set of rules the package now contains 3 files:
- out/main.js
- README.md
- package.json

The node_modules are not required for the plugin at runtime as the
minification should consume them and package it into the unified module.
This reduces the extension package from ~321 KiB to ~65 KiB.