vscode: address some OOB feedback
Remove the now extraneous `createDevPackage` script in favour of
`dev-package`.  Update the README.md accordingly.  Additionally, restore
the ignore rule for the test VSCode instance.

The gitignore is already configured to ignore vsix files, and vsce will
implicitly ignore the vsix files when packaging so we do not run the
risk of committing or distributing the vsix accidentally.