Use new way when building paths based on triple.
In TSC pr #229 and SPM pr #3576, we introduced a method named
`platformBuildPathComponent` to distinguish the valid platform path name
versus the valid target name, where one may have an embedded version
number and the other may not.

SourceKit test suite SwiftPMIntegrationTests currently fails. This is
presumably because this change may be required as it is one of two
uses of `tripleString`. The other uses `tripleString` as the value for
the `-target` field, so it is correct as written. The other -- as in
this commit -- builds a path.

This commit must be committed in concert with the aforementioned prs.