Don't report duplicate folding ranges
I’m not entirely sure when this started occurring but when retrieving folding ranges e.g. for
```
func foo() {
    print("someText")
}
```
We report the folding range of `"someText"` twice: Once as the body of the call to `print` and once as the body of the first argument. Each folding range on its own makes sense because you might want to collapse either all arguments or only a single argument. But it doesn’t make sense to report the same range twice.