ToolchainRegistryTests: don't do identity check for `localFileSystem` (#733)
We'd like to remove `AnyObject` requirement from `FileSystem`, which will make identity checks on `any FileSystem` impossible. These checks were an anti-pattern anyway, especially in tests, which should be flexible enough to support an arbitrary file system implementation. If certain behavior is not available on `FileSystem` we should modify that protocol instead.