Fix an issue that caused the wrong default toolchain to get selected
A previous commit in this PR changed the way the default toolchain was computed. That commit had a bug that it didn’t prefer the toolchain discovered from `SOURCEKIT_TOOLCHAIN_PATH` and relative to the `sourcekit-lsp` executable over the Xcode default toolchain. To fix that, record the reason why we found a toolchain and use that to prefer those toolchains over the Xcode default toolchain.