Append a suffix at then end of a chunk logged by `splitLongMultilineMessage`
`os_log` truncates trailing whitespace in a log message. If a chunk containing the file contents ends with a newline, this means that the newline will not get logged and thus the file contents read from the log aren't correct. Append an end marker to each chunk to prevent this trimming.