Add `SwiftDriver` dependency to `SKSwiftPMWorkspace`
In https://github.com/apple/swift-package-manager/pull/7258 we need to use `SwiftDriver` explicitly to make sure that `package` access modifier and `-package-name` support can be detected properly. Somehow this dependency needs to be propagate to everyone depending `SourceKitLSPAPI`, even though the dependency was added as `PRIVATE` in SwiftPM. Adding it as `PRIVATE` here too.