De-dupe locations returned from the defintions request
We might end up with duplicate locations when performing a definition request on eg. `MyStruct()` when no explicit initializer is declared. In this case we get two symbol infos, one for the declaration of the `MyStruct` type and one for the initializer, which is implicit and thus has the location of  the `MyStruct` declaration itself.

rdar://123036565
Fixes #1055