Only check if a document has been modified while loading diagnostics during publishDiagnostics
I was seeing some of these errors in VS Code, which uses the pull diagnostics request even though I couldn’t see any concurrent modifications to the document. Move the check up to `publishDiagnostics`, where it belongs, and log the version of the old and new document version so we can investiage what’s causing the issue.